+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - `view` command to show project label instead of app name
 - `rename` command to accept project label instead of app name
+- `update` command to accept project label instead of app name
 
 ### Removed